+++ Editor/Material/StandardLitShaderGUI.cs
@@ -146,6 +146,7 @@ namespace VividRP.Editor
 
         private const string AlphaTestKeyword = "_ALPHATEST_ON";
         private const string OpacityMapKeyword = "_OPACITYMAP";
+        private const string TransmissionMapKeyword = "_TRANSMISSIONMAP";
         private const string NormalMapKeyword = "_NORMALMAP";
         private const string MetallicGlossMapKeyword = "_METALLICSPECGLOSSMAP";
         private const string RoughnessMapKeyword = "_ROUGHNESSMAP";
@@ -271,6 +272,7 @@ namespace VividRP.Editor
         {
             CoreUtils.SetKeyword(material, AlphaTestKeyword, GetFloat(material, "_AlphaClip") > AlphaClipThreshold);
             CoreUtils.SetKeyword(material, OpacityMapKeyword, material.GetTexture("_OpacityMap") != null);
+            CoreUtils.SetKeyword(material, TransmissionMapKeyword, material.GetTexture("_TransmissionMap") != null);
             CoreUtils.SetKeyword(material, NormalMapKeyword, material.GetTexture("_BumpMap") != null);
             CoreUtils.SetKeyword(material, MetallicGlossMapKeyword, material.GetTexture("_MetallicGlossMap") != null);
             CoreUtils.SetKeyword(material, RoughnessMapKeyword, material.GetTexture("_RoughnessMap") != null);